+++ Magic/CHANGELOG.md
@@ -34,6 +34,7 @@
  - Add `until` parameter to Skip action, to skip sub-actions until some amount of time has passed
  - Add attributes: target_location_x, target_location_y, target_location_z
  - Add /mauto enable/disable commands to quickly turn on/off automata
+ - Add `give_to_caster` parameter to `Drop` action
 
 # 8.3.3